+++ engine/src/engine/search/negamax.rs
@@ -287,7 +287,7 @@ pub fn negamax(
     let mut best_move = None;
     let mut best_eval = Eval::MIN;
 
-    let mut moves = MovePicker::new(previous_best_move, Eval(good_tactical_see_bound()));
+    let mut moves = MovePicker::new(previous_best_move);
     let mut number_of_legal_moves = 0;
     let mut node_pv = PrincipalVariation::new();