+++ engine/src/engine/search/quiescence.rs
@@ -1,10 +1,7 @@
 use super::{MAX_SEARCH_DEPTH, SearchContext};
 use crate::{
     chess::game::Game,
-    engine::{
-        eval, eval::Eval, params::*, search::move_picker::MovePicker,
-        transposition_table::NodeBound,
-    },
+    engine::{eval, eval::Eval, search::move_picker::MovePicker, transposition_table::NodeBound},
 };
 
 pub fn quiescence(
@@ -92,7 +89,7 @@ pub fn quiescence(
     let mut best_eval = eval;
     let mut best_move = None;
 
-    let mut moves = MovePicker::new_loud(previous_best_move, Eval(qs_good_tactical_see_bound()));
+    let mut moves = MovePicker::new_loud(previous_best_move);
     while let Some(mv) = moves.next(game, ctx.tables, ctx.stack, plies) {
         ctx.tt.prefetch(game.approx_zobrist_after(mv));