+++ engine/src/tests/perft_tests.rs
@@ -1,9 +1,6 @@
 use crate::{
     chess::{fen::START_POS, game::Game, movegen, moves::MoveList, perft::perft},
-    engine::{
-        eval::Eval,
-        search::{SearchStack, move_picker::MovePicker, tables::Tables},
-    },
+    engine::search::{SearchStack, move_picker::MovePicker, tables::Tables},
 };
 
 const ENABLE_MOVEPICKER_PERFT: bool = false;
@@ -66,7 +63,7 @@ fn movepicker_perft(
     }
 
     let mut moves_at_this_node = Vec::new();
-    let mut movepicker = MovePicker::new(best_move, Eval(0));
+    let mut movepicker = MovePicker::new(best_move);
     while let Some(mv) = movepicker.next(game, tables, stack, depth) {
         game.make_move(mv);
         moves += movepicker_perft(depth - 1, game, tables, stack);